# Optimized parsing GET verb and version