### Before
```
                   Method |          Mean |     StdDev | Scaled | Scaled-StdDev |          RPS | Allocated |
------------------------- |-------------- |----------- |------- |-------------- |------------- |---------- |
           ParsePlaintext | 1,079.6836 ns |  8.1507 ns |   1.00 |          0.00 |   926,197.25 |     168 B |
  ParsePipelinedPlaintext |   877.8646 ns |  7.3689 ns |   0.81 |          0.01 | 1,139,127.77 |     168 B |
          ParseLiveAspNet | 3,826.8063 ns | 15.9470 ns |   3.54 |          0.03 |   261,314.51 |   2.02 kB |
 ParsePipelinedLiveAspNet | 3,828.8495 ns | 29.7095 ns |   3.55 |          0.04 |   261,175.06 |   2.02 kB |
             ParseUnicode | 6,870.2153 ns | 33.0606 ns |   6.36 |          0.06 |   145,555.85 |   3.41 kB |
    ParseUnicodePipelined | 6,944.8703 ns | 24.7683 ns |   6.43 |          0.05 |   143,991.17 |   3.41 kB |
```

### After
```
                   Method |          Mean |     StdDev | Scaled | Scaled-StdDev |          RPS | Allocated |
------------------------- |-------------- |----------- |------- |-------------- |------------- |---------- |
           ParsePlaintext | 1,061.8108 ns |  8.9208 ns |   1.00 |          0.00 |   941,787.34 |     168 B |
  ParsePipelinedPlaintext |   865.3721 ns |  6.0690 ns |   0.82 |          0.01 | 1,155,572.27 |     168 B |
          ParseLiveAspNet | 3,882.7252 ns | 85.4559 ns |   3.66 |          0.08 |   257,551.06 |   2.02 kB |
 ParsePipelinedLiveAspNet | 3,866.2318 ns | 21.6853 ns |   3.64 |          0.04 |   258,649.78 |   2.02 kB |
             ParseUnicode | 6,857.4925 ns | 67.3531 ns |   6.46 |          0.08 |   145,825.90 |   3.41 kB |
    ParseUnicodePipelined | 7,009.4501 ns | 39.9254 ns |   6.60 |          0.07 |   142,664.54 |   3.41 kB |
```

Wow, my home machine is faster than the one I have at work! 900K rps vs ~500K