@KrzysztofCwalina 2017-02-25
it's strange but the build system is not picking up the newest System.Slices from myget-corfxlab