@KrzysztofCwalina 2017-02-25
Please do not merge. I need to dig a bit more into this. Initially I saw a nice perf improvement, but I think it might be because I optimized Read in corfxlab. These Kestrel changes might not be improving anything or could be actually detrimental. 

I don't have time to do it today. I will finish it Monday.