@KrzysztofCwalina 2017-02-27
@davidfowl, I convinced myself there is a small improvement here. I think this should be merged.