@pakrym 2017-02-27
@KrzysztofCwalina you can write a benchmark just for these methods