@KrzysztofCwalina 2017-02-27
I did two things:
1. I ran the PlainText benchmark in more controlled eniroment, e.g. not browsing the web while running benchmarks, and the results were more consistent and in favod of the change.
2. I looked at the assembly. The after changes assembly is cleaner. The whole method is quite long, but here is the front of the method (the part that parses succesful GET):

###Before:
```assembly
          knownMethod = null;
00007FFCFC3A3482  push        rdi  
00007FFCFC3A3483  push        rsi  
00007FFCFC3A3484  push        rbp  
00007FFCFC3A3485  push        rbx  
00007FFCFC3A3486  sub         rsp,20h  
00007FFCFC3A348A  mov         rsi,rdx  
00007FFCFC3A348D  xor         eax,eax  
00007FFCFC3A348F  mov         qword ptr [rsi],rax  
00007FFCFC3A3492  cmp         dword ptr [rcx+10h],8  
00007FFCFC3A3496  jge         00007FFCFC3A34A5  
            {
                return false;
00007FFCFC3A3498  xor         eax,eax  
00007FFCFC3A349A  add         rsp,20h  
00007FFCFC3A349E  pop         rbx  
00007FFCFC3A349F  pop         rbp  
00007FFCFC3A34A0  pop         rsi  
00007FFCFC3A34A1  pop         rdi  
00007FFCFC3A34A2  pop         r14  
00007FFCFC3A34A4  ret  
            }

            ulong value = span.Read<ulong>();
00007FFCFC3A34A5  mov         rdi,qword ptr [rcx]  
00007FFCFC3A34A8  mov         rbx,qword ptr [rcx+8]  
00007FFCFC3A34AC  mov         eax,dword ptr [rcx+10h]  
00007FFCFC3A34AF  cmp         eax,8  
00007FFCFC3A34B2  jae         00007FFCFC3A34C1  
00007FFCFC3A34B4  call        00007FFCFC3A3188  
00007FFCFC3A34B9  mov         rcx,rax  
00007FFCFC3A34BC  call        00007FFD5BE7DC40  
00007FFCFC3A34C1  test        rdi,rdi  
00007FFCFC3A34C4  jne         00007FFCFC3A34C8  
00007FFCFC3A34C6  jmp         00007FFCFC3A34CF  
00007FFCFC3A34C8  lea         rcx,[rdi+8]  
00007FFCFC3A34CC  add         rbx,rcx  
00007FFCFC3A34CF  mov         rdi,qword ptr [rbx]  
            if ((value & _mask4Chars) == _httpGetMethodLong)
00007FFCFC3A34D2  mov         rcx,7FFCFC245640h  
00007FFCFC3A34DC  mov         edx,1  
00007FFCFC3A34E1  call        00007FFD5C0B8020  
00007FFCFC3A34E6  mov         rdx,rdi  
00007FFCFC3A34E9  and         rdx,qword ptr [7FFCFC245680h]  
00007FFCFC3A34F0  cmp         rdx,qword ptr [7FFCFC245678h]  
00007FFCFC3A34F7  jne         00007FFCFC3A351E  
            {
                knownMethod = Get;
00007FFCFC3A34F9  mov         rdx,27F313827A8h  
00007FFCFC3A3503  mov         rdx,qword ptr [rdx]  
00007FFCFC3A3506  mov         rcx,rsi  
00007FFCFC3A3509  call        00007FFD5C0B5AD0  
                return true;
00007FFCFC3A350E  mov         eax,1  
00007FFCFC3A3513  add         rsp,20h  
00007FFCFC3A3517  pop         rbx  
00007FFCFC3A3518  pop         rbp  
00007FFCFC3A3519  pop         rsi  
00007FFCFC3A351A  pop         rdi  
00007FFCFC3A351B  pop         r14  
00007FFCFC3A351D  ret  
```
###After:
```assembly
         if (span.TryRead<uint>(out var possiblyGet))
00007FFCFC3A35A2  push        rdi  
00007FFCFC3A35A3  push        rsi  
00007FFCFC3A35A4  push        rbp  
00007FFCFC3A35A5  push        rbx  
00007FFCFC3A35A6  sub         rsp,20h  
00007FFCFC3A35AA  mov         rsi,rcx  
00007FFCFC3A35AD  mov         rdi,rdx  
00007FFCFC3A35B0  mov         rcx,qword ptr [rsi]  
00007FFCFC3A35B3  mov         rdx,qword ptr [rsi+8]  
00007FFCFC3A35B7  mov         eax,dword ptr [rsi+10h]  
00007FFCFC3A35BA  and         eax,0FFFFFFFFh  
00007FFCFC3A35BF  cmp         rax,4  
00007FFCFC3A35C3  jl          00007FFCFC3A3616  
00007FFCFC3A35C5  test        rcx,rcx  
00007FFCFC3A35C8  jne         00007FFCFC3A35CC  
00007FFCFC3A35CA  jmp         00007FFCFC3A35D3  
00007FFCFC3A35CC  add         rcx,8  
00007FFCFC3A35D0  add         rdx,rcx  
00007FFCFC3A35D3  mov         ebx,dword ptr [rdx]  
            {
                if (possiblyGet == _httpGetMethodInt)
00007FFCFC3A35D5  mov         rcx,7FFCFC245640h  
00007FFCFC3A35DF  mov         edx,1  
00007FFCFC3A35E4  call        00007FFD5C0B8020  
00007FFCFC3A35E9  cmp         ebx,dword ptr [7FFCFC245688h]  
00007FFCFC3A35EF  jne         00007FFCFC3A3616  
                {
                    knownMethod = Get;
00007FFCFC3A35F1  mov         rdx,27F313827A8h  
00007FFCFC3A35FB  mov         rdx,qword ptr [rdx]  
00007FFCFC3A35FE  mov         rcx,rdi  
00007FFCFC3A3601  call        00007FFD5C0B5AD0  
                    return true;
00007FFCFC3A3606  mov         eax,1  
00007FFCFC3A360B  add         rsp,20h  
00007FFCFC3A360F  pop         rbx  
00007FFCFC3A3610  pop         rbp  
00007FFCFC3A3611  pop         rsi  
00007FFCFC3A3612  pop         rdi  
00007FFCFC3A3613  pop         r14  
00007FFCFC3A3615  ret  
```