@KrzysztofCwalina 2017-02-27
If you still want me to write benchmarks, I can, but to me the codegen difference is pretty convincing. Let me know.