@benaadams 2017-02-27
Looking at the asm, the current would work better if `_maskXChars` were consts rather than readonly static conversions from byte arrays (so memory load on use)