@cesarbs 2017-02-27
@KrzysztofCwalina Please add the benchmark as requested by @pakrym. While the shorter code is good evidence, we need an actual measurement. E2E benchmarking doesn't catch this change (too much noise).