@KrzysztofCwalina 2017-02-27
@benaadams, I changed _httpGetMethodInt to a const and the "if (possiblyGet == _httpGetMethodInt)" block codegened to:
```assembly
00007FFCFCF53473  mov         edx,dword ptr [rax]  
            {
                if (possiblyGet == _httpGetMethodInt)
00007FFCFCF53475  cmp         edx,20544547h  
00007FFCFCF5347B  jne         00007FFCFCF534B6  
                {
                    knownMethod = Get;
```

I update the PR to use const