@benaadams 2017-02-27
Could change the `ulong`s to consts also ;-)