@KrzysztofCwalina 2017-02-27
Here are the microbenchmark results:

``` ini

BenchmarkDotNet=v0.10.1, OS=Microsoft Windows NT 6.2.9200.0
Processor=Intel(R) Core(TM) i7-6700K CPU 4.00GHz, ProcessorCount=8
Frequency=3914061 Hz, Resolution=255.4891 ns, Timer=TSC
  [Host]     : Clr 4.0.30319.42000, 64bit RyuJIT-v4.6.1586.0
  DefaultJob : Clr 4.0.30319.42000, 64bit RyuJIT-v4.6.1586.0

Allocated=0 B  

```
 Method |       Mean |    StdDev |
------- |----------- |---------- |
  After | 28.5595 us | 0.0070 us |
 Before | 36.1777 us | 0.1141 us |
```