@KrzysztofCwalina 2017-02-27
@benaadams, I will leave it up to the Kestrel guys to decided if they want consts for all the ulongs. They are not on super critical path, and I am not sure if we want to tradeoff maintainability for a couple of instructions.