@arespr 2017-02-27
I have found a function that creates key for lookup into array _knownMethods with extended length to 17.
It's twice faster than foreach (var x in _knownMethods) in microbenchmarks:

https://github.com/arespr/ConsoleApp1/blob/09ac0c1635914eb55b546c1c8b8fcba75f02e4ec/ConsoleApp1/KnownMethods.cs